@@ -25,31 +25,31 @@
     ]
   },
   "dependencies": {
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "styled-components": "5.3.3"
+    "react": "18.1.0",
+    "react-dom": "18.1.0",
+    "styled-components": "5.3.5"
   },
   "devDependencies": {
-    "@types/node": "17.0.22",
-    "@types/react": "17.0.41",
-    "@types/react-dom": "17.0.14",
-    "@types/styled-components": "5.1.24",
-    "@typescript-eslint/eslint-plugin": "5.16.0",
-    "@vitejs/plugin-react": "1.2.0",
-    "esbuild": "0.14.27",
-    "eslint": "8.11.0",
-    "eslint-import-resolver-typescript": "2.7.0",
-    "eslint-plugin-import": "2.25.4",
-    "eslint-plugin-react-hooks": "4.3.0",
+    "@types/node": "17.0.31",
+    "@types/react": "18.0.8",
+    "@types/react-dom": "18.0.3",
+    "@types/styled-components": "5.1.25",
+    "@typescript-eslint/eslint-plugin": "5.22.0",
+    "@vitejs/plugin-react": "1.3.2",
+    "esbuild": "0.14.38",
+    "eslint": "8.14.0",
+    "eslint-import-resolver-typescript": "2.7.1",
+    "eslint-plugin-import": "2.26.0",
+    "eslint-plugin-react-hooks": "4.5.0",
     "eslint-plugin-unused-imports": "2.0.0",
     "husky": "7.0.4",
-    "lint-staged": "12.3.7",
-    "sass": "1.49.9",
+    "lint-staged": "12.4.1",
+    "sass": "1.51.0",
     "shx": "0.3.4",
-    "tslib": "2.3.1",
-    "typescript": "4.6.2",
+    "tslib": "2.4.0",
+    "typescript": "4.6.4",
     "uglifycss": "0.0.29",
-    "vite": "2.8.6",
+    "vite": "2.9.8",
     "vite-tsconfig-paths": "3.4.1"
   }
 }
