@@ -1,12 +1,13 @@
 import * as React from "react";
-import * as ReactDOM from "react-dom";
+import * as ReactDOM from "react-dom/client";
 import App from "./App";
 
-ReactDOM.render(
+const root = ReactDOM.createRoot(document.getElementById("App")!)
+
+root.render(
 	<React.StrictMode>
 		<App />
-	</React.StrictMode>,
-	document.getElementById("App")
+	</React.StrictMode>
 );
 
 // eslint-disable-next-line import/no-unused-modules
